90.10d_RAG_OnTheFly/requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-community
PyPDF2
python-docx
python-pptx
faiss-cpu
sentence-transformers
beautifulsoup4
html2text
